New pattern MVVM
