+++ app/src/main/AndroidManifest.xml
@@ -4,6 +4,7 @@
     package="com.devchernikova.firstapp">
 
     <application
+        android:name=".App"
         android:allowBackup="true"
         android:icon="@drawable/ic_baseline_favorite_24"
         android:label="@string/app_name"
@@ -11,7 +12,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".DetailsFragment"
+            android:name=".view.fragments.DetailsFragment"
             android:exported="true"
             tools:ignore="Instantiatable" />
         <activity
