+++ app/src/main/java/com/devchernikova/firstapp/MainActivity.kt
@@ -5,6 +5,8 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import com.devchernikova.firstapp.databinding.ActivityMainBinding
+import com.devchernikova.firstapp.domain.Film
+import com.devchernikova.firstapp.view.fragments.*
 
 
 class MainActivity : AppCompatActivity() {
