+++ app/src/main/java/com/devchernikova/firstapp/domain/Interactor.kt
@@ -1,0 +1,7 @@
+package com.devchernikova.firstapp.domain
+
+import com.devchernikova.firstapp.data.MainRepository
+
+class Interactor(val repo: MainRepository) {
+    fun getFilmsDB(): List<Film> = repo.filmsDataBase
+}
