+++ app/src/main/java/com/devchernikova/firstapp/view/customviews/RatingDonutView.kt
@@ -1,9 +1,10 @@
-package com.devchernikova.firstapp
+package com.devchernikova.firstapp.view.customviews
 
 import android.content.Context
 import android.graphics.*
 import android.util.AttributeSet
 import android.view.View
+import com.devchernikova.firstapp.R
 
 class RatingDonutView @JvmOverloads constructor(context: Context, attributeSet: AttributeSet? = null) : View(context, attributeSet) {
     //Овал для рисования сегментов прогресс бара
