+++ app/src/main/java/com/devchernikova/firstapp/view/fragments/WatchLaterFragment.kt
@@ -1,12 +1,11 @@
-package com.devchernikova.firstapp
+package com.devchernikova.firstapp.view.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
-import com.devchernikova.firstapp.databinding.FragmentSelectionsBinding
+import com.devchernikova.firstapp.R
 import com.devchernikova.firstapp.databinding.FragmentWatchLaterBinding
+import com.devchernikova.firstapp.utils.AnimationHelper
 
 
 class WatchLaterFragment : Fragment(R.layout.fragment_watch_later) {
