+++ app/src/main/java/com/devchernikova/firstapp/viewmodel/HomeFragmentViewModel.kt
@@ -1,0 +1,18 @@
+package com.devchernikova.firstapp.viewmodel
+
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import com.devchernikova.firstapp.App
+import com.devchernikova.firstapp.domain.Film
+import com.devchernikova.firstapp.domain.Interactor
+
+class HomeFragmentViewModel : ViewModel() {
+    val filmsListLiveData:  MutableLiveData<List<Film>> = MutableLiveData()
+    //Инициализируем интерактор
+    private var interactor: Interactor = App.instance.interactor
+
+    init {
+        val films = interactor.getFilmsDB()
+        filmsListLiveData.postValue(films)
+    }
+}
