+++ app/src/main/res/layout/fragment_watch_later.xml
@@ -7,6 +7,6 @@
     android:background="@drawable/shapewatchlater"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WatchLaterFragment">
+    tools:context=".view.fragments.WatchLaterFragment">
 
 </FrameLayout>
